@@ -3,120 +3,47 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@types/immutable": "^3.8.7",
+    "classnames": "^2.2.6",
+    "connected-react-router": "^6.6.1",
+    "history": "^4.10.1",
+    "immutable": "^4.0.0-rc.12",
+    "jest": "^25.1.0",
+    "lodash": "^4.17.15",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "react-redux": "^7.1.3",
+    "react-router-dom": "^5.1.2",
+    "react-router-redux": "^4.0.8",
+    "react-scripts": "^3.3.0",
+    "redux": "^4.0.5",
+    "redux-devtools": "^3.5.0",
+    "redux-devtools-dock-monitor": "^1.1.3",
+    "redux-devtools-log-monitor": "^1.4.0"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test"
+  },
+  "devDependencies": {
     "@types/classnames": "^2.2.3",
     "@types/history": "^4.6.0",
+    "@types/jest": "^21.1.1",
     "@types/lodash": "^4.14.77",
+    "@types/node": "^8.0.31",
+    "@types/react": "^16.0.7",
+    "@types/react-dom": "^16.0.1",
     "@types/react-redux": "^5.0.10",
-    "@types/react-router": "^4.0.15",
-    "@types/react-router-dom": "^4.0.8",
-    "@types/react-router-redux": "^5.0.8",
+    "@types/react-router": "^5.1.4",
+    "@types/react-router-dom": "^5.1.3",
+    "@types/react-router-redux": "^5.0.18",
     "@types/redux": "^3.6.0",
     "@types/redux-devtools": "^3.0.40",
     "@types/redux-devtools-dock-monitor": "^1.1.31",
     "@types/redux-devtools-log-monitor": "^1.0.32",
-    "autoprefixer": "7.1.2",
-    "case-sensitive-paths-webpack-plugin": "2.1.1",
-    "chalk": "1.1.3",
-    "classnames": "^2.2.5",
-    "copy-webpack-plugin": "^4.1.1",
-    "css-loader": "0.28.4",
-    "dotenv": "4.0.0",
-    "extract-text-webpack-plugin": "3.0.0",
-    "file-loader": "0.11.2",
-    "fs-extra": "3.0.1",
-    "gulp": "^3.9.1",
-    "gulp-concat": "^2.6.1",
-    "history": "^4.7.2",
-    "html-webpack-plugin": "2.29.0",
-    "immutable": "4.0.0-rc.7",
-    "jest": "20.0.4",
-    "lodash": "^4.17.4",
-    "midi": "github:mudcube/MIDI.js",
-    "node-sass": "^4.5.3",
-    "object-assign": "4.1.1",
-    "postcss-flexbugs-fixes": "3.2.0",
-    "postcss-loader": "2.0.6",
-    "promise": "8.0.1",
-    "react": "^16.0.0",
-    "react-dev-utils": "^4.0.1",
-    "react-dom": "^16.0.0",
-    "react-redux": "^5.0.6",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.2.2",
-    "react-router-redux": "^5.0.0-alpha.6",
-    "redux": "^3.7.2",
-    "redux-devtools": "^3.4.0",
-    "redux-devtools-dock-monitor": "^1.1.2",
-    "redux-devtools-log-monitor": "^1.3.0",
-    "sass-loader": "^6.0.6",
-    "source-map-loader": "^0.2.1",
-    "style-loader": "0.18.2",
-    "sw-precache-webpack-plugin": "0.11.4",
-    "ts-jest": "^20.0.7",
-    "ts-loader": "^2.2.1",
-    "tslint": "^5.2.0",
-    "tslint-loader": "^3.5.3",
-    "tslint-react": "^3.0.0",
-    "typescript": "~2.5.0",
-    "url-loader": "0.5.9",
-    "webpack": "3.5.1",
-    "webpack-dev-server": "2.7.1",
-    "webpack-manifest-plugin": "1.2.1",
-    "whatwg-fetch": "2.0.3"
-  },
-  "scripts": {
-    "start": "node scripts/start.js",
-    "build": "node scripts/build.js",
-    "test": "node scripts/test.js --env=jsdom"
-  },
-  "devDependencies": {
-    "@types/jest": "^21.1.1",
-    "@types/node": "^8.0.31",
-    "@types/react": "^16.0.7",
-    "@types/react-dom": "^16.0.1"
-  },
-  "jest": {
-    "mapCoverage": true,
-    "collectCoverageFrom": [
-      "src/**/*.{js,jsx,ts,tsx}"
-    ],
-    "setupFiles": [
-      "<rootDir>/config/polyfills.js"
-    ],
-    "testMatch": [
-      "<rootDir>/src/**/__tests__/**/*.ts?(x)",
-      "<rootDir>/src/**/?(*.)(spec|test).ts?(x)"
-    ],
-    "testEnvironment": "node",
-    "testURL": "http://localhost",
-    "transform": {
-      "^.+\\.css$": "<rootDir>/config/jest/cssTransform.js",
-      "^.+\\.tsx?$": "<rootDir>/config/jest/typescriptTransform.js",
-      "^(?!.*\\.(js|jsx|css|json)$)": "<rootDir>/config/jest/fileTransform.js"
-    },
-    "transformIgnorePatterns": [
-      "[/\\\\]node_modules[/\\\\].+\\.(js|jsx|ts|tsx)$"
-    ],
-    "moduleNameMapper": {
-      "^react-native$": "react-native-web"
-    },
-    "moduleFileExtensions": [
-      "web.ts",
-      "ts",
-      "web.tsx",
-      "tsx",
-      "web.js",
-      "js",
-      "web.jsx",
-      "jsx",
-      "json",
-      "node"
-    ],
-    "globals": {
-      "ts-jest": {
-        "tsConfigFile": "/Users/danielchao/code/chord-substitutioner/tsconfig.test.json"
-      }
-    }
+    "node-sass": "^4.13.1",
+    "typescript": "^3.7.5"
   },
   "babel": {
     "presets": [
