@@ -1,5 +1,5 @@
 import { SoundAction } from "./soundActions";
 import { LoaderAction } from "./loaderActions";
-import { RouterAction, LocationChangeAction } from "react-router-redux";
+import { RouterAction, LocationChangeAction } from "connected-react-router";
 
 export type AppAction = SoundAction | LoaderAction | RouterAction | LocationChangeAction;
