@@ -1,4 +1,4 @@
-import * as React from "react";
+import React from "react";
 import { createDevTools } from "redux-devtools";
 import DockMonitor from "redux-devtools-dock-monitor";
 import LogMonitor from "redux-devtools-log-monitor";
