@@ -1,4 +1,4 @@
-import * as React from "react";
+import React from "react";
 import { IAppProps } from "../containers/App";
 import { midiLoaded } from "../actions/loaderActions";
 import { replace } from "react-router-redux";
@@ -8,6 +8,7 @@ import connectedComponent from "../util/connectedComponent";
 export default class Loader extends React.Component<IAppProps> {
 
   componentWillMount () {
+    debugger;
     MIDI.loadPlugin({
       instruments: ["acoustic_grand_piano", "choir_aahs", "string_ensemble_1", "synth_choir", "woodblock", "xylophone"],
       soundfontUrl: "soundfonts/",
