@@ -1,8 +1,9 @@
 import { combineReducers } from "redux";
 import soundReducer, { ISoundState } from "./soundReducer";
-import { routerReducer, RouterState } from "react-router-redux";
+import { connectRouter, RouterState } from "connected-react-router";
 import loaderReducer, { ILoaderState } from "./loaderReducer";
 import songReducer, { ISongState } from "./songReducers";
+import { History } from "history";
 
 export interface IAppState {
   sound: ISoundState;
@@ -11,9 +12,11 @@ export interface IAppState {
   song: ISongState;
 }
 
-export default combineReducers({
+const createReducer = (history: History) => combineReducers({
   sound: soundReducer,
-  router: routerReducer,
+  router: connectRouter(history),
   loader: loaderReducer,
   song: songReducer
 });
+
+export default createReducer;
